+++ crates/cargo-contract/src/cmd/extrinsics/upload.rs
@@ -15,6 +15,7 @@
 // along with cargo-contract.  If not, see <http://www.gnu.org/licenses/>.
 
 use super::{
+    display_dry_run_result_warning,
     runtime_api::api::{
         self,
         runtime_types::pallet_contracts::wasm::Determinism,
@@ -78,7 +79,7 @@ impl UploadCommand {
             let url = self.extrinsic_opts.url_to_string();
             let client = OnlineClient::from_url(url.clone()).await?;
 
-            if self.extrinsic_opts.dry_run {
+            if !self.extrinsic_opts.execute {
                 match self.upload_code_rpc(code, &client, &signer).await? {
                     Ok(result) => {
                         let upload_result = UploadDryRunResult {
@@ -90,6 +91,7 @@ impl UploadCommand {
                             println!("{}", upload_result.to_json()?);
                         } else {
                             upload_result.print();
+                            display_dry_run_result_warning("upload");
                         }
                     }
                     Err(err) => {
@@ -102,7 +104,6 @@ impl UploadCommand {
                         }
                     }
                 }
-                Ok(())
             } else if let Some(code_stored) =
                 self.upload_code(&client, code, &signer).await?
             {
@@ -114,14 +115,14 @@ impl UploadCommand {
                 } else {
                     upload_result.print();
                 }
-                Ok(())
             } else {
                 let code_hash = hex::encode(code_hash);
-                Err(anyhow::anyhow!(
+                return Err(anyhow::anyhow!(
                     "This contract has already been uploaded with code hash: 0x{code_hash}"
                 )
                 .into())
             }
+            Ok(())
         })
     }